refactor: move isDirect logic outside of edit button
