+++ package/src/components/ChannelDetails/hooks/useCanEdit.ts
@@ -1,0 +1,9 @@
+import type { Channel } from 'stream-chat';
+
+import { useChannelOwnCapabilities } from '../../../hooks/useChannelOwnCapabilities';
+
+/**
+ * Whether the current user can update (edit) the channel.
+ */
+export const useCanEdit = (channel?: Channel) =>
+  useChannelOwnCapabilities(channel)?.includes('update-channel') ?? false;
